% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_bart_y1}
\alias{cpp_bart_y1}
\title{Multiply a number by two}
\usage{
cpp_bart_y1(new_xroot, new_yroot, new_nd, new_burn, new_m, new_nu, new_kfac,
  new_nmissing, new_xmissroot, new_bistart, new_vartyperoot, new_zroot,
  new_beta, new_vroot, new_ffname, new_lambda, new_type)
}
\arguments{
\item{x}{many single integers.}
}
\description{
Multiply a number by two
}

