% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{datatype}
\alias{datatype}
\title{Data values for dataypes
A dummy dataset to run the function Seq BART on.}
\format{This dataframe has 4 rows}
\usage{
datatype
}
\description{
Data values for dataypes
A dummy dataset to run the function Seq BART on.
}
\keyword{datasets}

