% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{xx}
\alias{xx}
\title{Data values for variable xx for the Seq BART function
A dummy dataset to run the function Seq BART on.}
\format{This dataframe has 150 rows and 4 columns having missing values at random.}
\usage{
xx
}
\description{
Data values for variable xx for the Seq BART function
A dummy dataset to run the function Seq BART on.
}
\keyword{datasets}

