% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{y1}
\alias{y1}
\title{Data values for variable y (continuous) for the Seq BART function
A dummy dataset to run the function Seq BART on.}
\format{This dataframe has 150 values.}
\usage{
y1
}
\description{
Data values for variable y (continuous) for the Seq BART function
A dummy dataset to run the function Seq BART on.
}
\keyword{datasets}

