% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunctionBART.R
\name{seqBART}
\alias{seqBART}
\title{Sequential Bayesian Additive Regression Trees Model}
\usage{
seqBART(x, x.type, y.type = 0, y = NA, numimpute = 5, numskip = 199,
  burn = 1000, sigest = NA, seed = NA)
}
\arguments{
\item{x}{Dataset of covariate matrix with missing values (NAs).}

\item{x.type}{A vector indicating the type of covariates (0=continuous, 1=binary).}

\item{y.type}{0 for no response, 1 for continuous response using linear regression for imputation, 2 for binary response using logistic regression for imputation, 3 for continuous response using BART for imputation, 4 for binary response using BART probit for imputation.}

\item{y}{Response (fully observed).  Default is NA ( which is the case for type = 0); For other types, data must be provided.}

\item{numimpute}{Number of Imputed Datasets, default = 5}

\item{numskip}{Number of iterations skipped, default = 199}

\item{burn}{Number of iterations for burn-in, default value = 1000}

\item{sigest}{For continuous variable, an estimate of error variance, sigma^2, used to calibrate an inverse-chi-squared prior used on that parameter. If not supplied, the least-squares estimate is derived instead. See sigquant for more information. Not applicable when variable is binary. Default value is NA.}

\item{seed}{is the value that will used to generate the distributions and draws with. Default value is NA.}
}
\value{
Imputed Dataset Values named as 'imputed#'.
}
\description{
A flexible Bayesian nonparametric model that is used as imputation tool for missing covariates.
}
\examples{
{
# Prepare the Input Dataset
n=10
p=4
set.seed(12345)
x<-matrix(NA,n,p)
varm1<-matrix(0.8,p,p)
diag(varm1)<-1
library(MASS)
x<-mvrnorm(n,rep(0,p),varm1)
x[x[,3]>0,3]<-1
x[x[,3]<=0,3]<-0

y<-rowMeans(x)+rnorm(n) # y argument is required for y.type = 1/2/3/4

for(i in 2:4){
  ran<-runif(n)
  x[ran<0.1,i]=NA
}

x.type=c(0,0,1,0)
y.type=1
impute<-seqBART(x=x, y=y, x.type=x.type, y.type=y.type, seed= 12345) # call the function

}

}
